Enhance asset and user management: update asset table schema, improve DTO validation, implement pagination in asset retrieval, and refactor error handling

+++ internal/repository/user_repository.go
@@ -8,5 +8,5 @@ import (
 
 type UserRepositoryInterface interface {
 	Create(ctx context.Context, user *model.User) error
-	Delete(ctx context.Context, userId int) error
+	Delete(ctx context.Context, userId int64) error
 }
